Make effects and operations feel like builtin primitives
Reimplements effects and operations in C using `JanetAbstractType`.
This depends on a patched version of Janet that adds `callable` to
`JanetAbstractType`, which is unfortunate.  But, the outcome is worth
it.  Now operations look and act just like normal functions, effects
act like a dictionary of operation names to operations, and both effects
and operations print in a way that is less confusing than before.

Hopefully something like the patch to Janet can eventually be
incorporated upstream.